other/mxshop_srvs/user_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
protobuf
loguru
python-consul2